Create hooks folder inside movieScreen folder, move useMovies to that folder
